+++ public/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2026-03-26
+
+- **[chore]** **Countries:** Improved filtering, sorting and property searching
+- **[chore]** **General:** Updated autocomplete behavior
+
 ## 2026-03-23
 
 - **[chore]** **Countries:** Added new properties for filtering with property search